qa-agent/agent-api/requirements.txt
# QA Agent API Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Kubernetes client
kubernetes==29.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# YAML parsing
pyyaml==6.0.1

# Logging and monitoring
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
python-dotenv==1.0.0
